rename ops to operations
