+++ qutip/qip/algorithms/qft.py
@@ -38,7 +38,7 @@ This module provides the circuit implementation for Quantum Fourier Transform.
 import numpy as np
 import scipy.sparse as sp
 from qutip.qobj import *
-from qutip.qip.ops.gates import snot, cphase, swap
+from qutip.qip.operations.gates import snot, cphase, swap
 from qutip.qip.circuit import QubitCircuit
 from qutip.qobj import Qobj
