+++ qutip/qip/gates.py
@@ -32,8 +32,8 @@
 ###############################################################################
 import warnings
 
-from qutip.qip.ops.gates import *
+from qutip.qip.operations.gates import *
 warnings.warn(
     "Importation from qutip.qip.gates is deprecated."
-    "Please use e.g.\n from qutip.qip.ops import cnot",
-    DeprecationWarning)
+    "Please use e.g.\n from qutip.qip.operations import cnot",
+    DeprecationWarning, stacklevel=1)
