+++ qutip/qip/operations/__init__.py
@@ -30,7 +30,7 @@
 #    (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 #    OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ###############################################################################
-from qutip.qip.ops.gates import (
+from qutip.qip.operations.gates import (
     rx, ry, rz, sqrtnot, snot, phasegate, qrot,
     cphase, cnot,
     csign, berkeley, swapalpha, swap, iswap, sqrtswap,
