+++ qutip/qip/operations/gates.py
