+++ qutip/tests/test_cqed.py
@@ -35,7 +35,7 @@ import warnings
 import numpy as np
 from numpy.testing import assert_, run_module_suite, assert_allclose
 
-from qutip.qip.ops.gates import gate_sequence_product
+from qutip.qip.operations.gates import gate_sequence_product
 from qutip.qip.circuit import QubitCircuit
 from qutip.qip.device.cqed import DispersivecQED
 from qutip.random_objects import rand_ket
