+++ qutip/tests/test_entropy.py
@@ -39,7 +39,7 @@ from numpy.testing import assert_, assert_equal, run_module_suite
 from qutip import (basis, ket2dm, entropy_vn, entropy_linear, rand_ket,
                    rand_dm, tensor, concurrence, entropy_mutual, ptrace,
                    entropy_conditional, entangling_power)
-from qutip.qip.ops.gates import cnot, iswap, swap, berkeley, sqrtswap, swapalpha
+from qutip.qip.operations.gates import cnot, iswap, swap, berkeley, sqrtswap, swapalpha
 
 
 def test_EntropyVN():
